thesaurus factory
