+++ extension/aodn-classifier/src/main/java/au/org/emii/classifier/AodnThesaurusFactory.java
@@ -1,0 +1,28 @@
+package au.org.emii.classifier;
+
+import org.apache.log4j.Logger;
+import org.fao.geonet.kernel.Thesaurus;
+import org.fao.geonet.kernel.ThesaurusFinder;
+
+public class AodnThesaurusFactory {
+
+    private static Logger logger = Logger.getLogger(AodnThesaurusFactory.class);
+    private static ThesaurusFinder thesaurusFinder;
+
+    public AodnThesaurusFactory(ThesaurusFinder thesaurusFinder) {
+        this.thesaurusFinder = thesaurusFinder;
+    }
+
+    public IAodnThesaurus findThesaurus(String scheme) {
+
+        Thesaurus thesaurus = thesaurusFinder.getThesaurusByConceptScheme(scheme);
+        if (thesaurus == null) {
+            logger.info(String.format("Thesaurus not found for scheme='%s'", scheme));
+            return new NullThesaurus();
+        } else {
+            return new AodnTermsThesaurus(thesaurus);
+        }
+
+    }
+
+}
