+++ extension/aodn-classifier/src/main/java/au/org/emii/classifier/LabelClassifier.java
@@ -28,23 +28,12 @@ public class LabelClassifier implements Classifier {
         this.indexKey = indexKey;
     }
 
-    public IAodnThesaurus findThesaurus(String scheme) {
-
-        Thesaurus thesaurus = thesaurusFinder.getThesaurusByConceptScheme(scheme);
-        if (thesaurus == null) {
-            logger.info(String.format("Thesaurus not found for scheme='%s'", scheme));
-            return new NullThesaurus();
-        } else {
-            return new AodnTermsThesaurus(thesaurus);
-        }
-
-    }
-
     @Override
     public List<CategoryPath> classify(String value) {
 
-        IAodnThesaurus vocabularyThesaurus = findThesaurus(vocabularyScheme);
-        IAodnThesaurus classificationThesaurus = findThesaurus(classificationScheme);
+        AodnThesaurusFactory thesaurusFactory = new AodnThesaurusFactory(thesaurusFinder);
+        IAodnThesaurus vocabularyThesaurus = thesaurusFactory.findThesaurus(vocabularyScheme);
+        IAodnThesaurus classificationThesaurus = thesaurusFactory.findThesaurus(classificationScheme);
         AodnTermClassifier termClassifier = new AodnTermClassifier(vocabularyThesaurus, classificationThesaurus);
 
         List<AodnTerm> matchingTerms = vocabularyThesaurus.getTermWithLabel(value);
